Subscribers selector added to page
